requirements.txt
mutagen~=1.46.0
requests==2.31.0
requests-cache~=1.0.1
python-dotenv~=1.0.0
tqdm~=4.65.0
Pillow~=9.5.0
titlecase~=2.3
python-dateutil~=2.8.2
pyyaml~=6.0
pytest~=7.3.1
pytz~=2023.3
xmltodict~=0.13.0
lxml~=4.9.2
pyfiglet~=0.8post1